+++ objc2/src/foundation/object.rs
@@ -27,7 +27,7 @@ unsafe impl ClassType for NSObject {
 extern_methods!(
     unsafe impl NSObject {
         pub fn new() -> Id<Self, Owned> {
-            unsafe { msg_send_id![Self::class(), new].unwrap() }
+            unsafe { msg_send_id![Self::class(), new] }
         }
 
         fn is_kind_of_inner(&self, cls: &Class) -> bool {