+++ test-assembly/crates/test_msg_send_id/lib.rs
@@ -4,8 +4,23 @@ use objc2::rc::{Allocated, Id, Shared};
 use objc2::runtime::{Class, Object, Sel};
 
 #[no_mangle]
-unsafe fn handle_alloc(obj: &Class, sel: Sel) -> Option<Id<Allocated<Object>, Shared>> {
-    <RetainSemantics<false, true, false, false>>::send_message_id(obj, sel, ())
+unsafe fn handle_new(cls: &Class, sel: Sel) -> Option<Id<Object, Shared>> {
+    <RetainSemantics<true, false, false, false>>::send_message_id(cls, sel, ())
+}
+
+#[no_mangle]
+unsafe fn handle_new_fallible(cls: &Class, sel: Sel) -> Id<Object, Shared> {
+    <RetainSemantics<true, false, false, false>>::send_message_id(cls, sel, ())
+}
+
+#[no_mangle]
+unsafe fn handle_alloc(cls: &Class, sel: Sel) -> Option<Id<Allocated<Object>, Shared>> {
+    <RetainSemantics<false, true, false, false>>::send_message_id(cls, sel, ())
+}
+
+#[no_mangle]
+unsafe fn handle_alloc_fallible(cls: &Class, sel: Sel) -> Id<Allocated<Object>, Shared> {
+    <RetainSemantics<false, true, false, false>>::send_message_id(cls, sel, ())
 }
 
 #[no_mangle]
@@ -17,24 +32,32 @@ unsafe fn handle_init(
 }
 
 #[no_mangle]
-unsafe fn handle_alloc_init(obj: &Class, sel1: Sel, sel2: Sel) -> Option<Id<Object, Shared>> {
-    let obj = <RetainSemantics<false, true, false, false>>::send_message_id(obj, sel1, ());
+unsafe fn handle_init_fallible(
+    obj: Option<Id<Allocated<Object>, Shared>>,
+    sel: Sel,
+) -> Id<Object, Shared> {
+    <RetainSemantics<false, false, true, false>>::send_message_id(obj, sel, ())
+}
+
+#[no_mangle]
+unsafe fn handle_alloc_init(cls: &Class, sel1: Sel, sel2: Sel) -> Option<Id<Object, Shared>> {
+    let obj = <RetainSemantics<false, true, false, false>>::send_message_id(cls, sel1, ());
     <RetainSemantics<false, false, true, false>>::send_message_id(obj, sel2, ())
 }
 
 #[no_mangle]
 unsafe fn handle_alloc_release(cls: &Class, sel: Sel) {
-    let _obj: Id<Allocated<Object>, Shared> =
-        <RetainSemantics<false, true, false, false>>::send_message_id(cls, sel, ())
-            .unwrap_unchecked();
+    let obj: Option<Id<Allocated<Object>, Shared>> =
+        <RetainSemantics<false, true, false, false>>::send_message_id(cls, sel, ());
+    let _obj = obj.unwrap_unchecked();
 }
 
 #[no_mangle]
 unsafe fn handle_alloc_init_release(cls: &Class, sel1: Sel, sel2: Sel) {
     let obj = <RetainSemantics<false, true, false, false>>::send_message_id(cls, sel1, ());
-    let _obj: Id<Object, Shared> =
-        <RetainSemantics<false, false, true, false>>::send_message_id(obj, sel2, ())
-            .unwrap_unchecked();
+    let obj: Option<Id<Object, Shared>> =
+        <RetainSemantics<false, false, true, false>>::send_message_id(obj, sel2, ());
+    let _obj = obj.unwrap_unchecked();
 }
 
 #[no_mangle]
@@ -43,6 +66,16 @@ unsafe fn handle_copy(obj: &Object, sel: Sel) -> Option<Id<Object, Shared>> {
 }
 
 #[no_mangle]
+unsafe fn handle_copy_fallible(obj: &Object, sel: Sel) -> Id<Object, Shared> {
+    <RetainSemantics<false, false, false, true>>::send_message_id(obj, sel, ())
+}
+
+#[no_mangle]
 unsafe fn handle_autoreleased(obj: &Object, sel: Sel) -> Option<Id<Object, Shared>> {
+    <RetainSemantics<false, false, false, false>>::send_message_id(obj, sel, ())
+}
+
+#[no_mangle]
+unsafe fn handle_autoreleased_fallible(obj: &Object, sel: Sel) -> Id<Object, Shared> {
     <RetainSemantics<false, false, false, false>>::send_message_id(obj, sel, ())
 }